+++ packages/slog_basic/lib/src/interceptors.dart
@@ -4,20 +4,23 @@ import 'dart:convert';
 import 'package:slog_core/slog_core.dart';
 import 'package:stack_trace/stack_trace.dart';
 
+///Prints the current Log request in a JSON format
 class PrintInterceptor extends SLogInterceptor {
+  ///Prints the current Log request in a JSON format
   PrintInterceptor({this.shouldPrint = true});
 
+  ///Will only print if `true`
   final bool shouldPrint;
 
   @override
   FutureOr<LogRequest> onRequestEvent(LogRequest request) {
     if (shouldPrint) {
-      print(getPrettyLogRequest(request));
+      print(_getPrettyLogRequest(request));
     }
     return request;
   }
 
-  String getPrettyLogRequest(LogRequest request) {
+  String _getPrettyLogRequest(LogRequest request) {
     final prettyJson = _getPrettyJSONString(request.toMap());
 
     final out = '''