+++ packages/slog_basic/test/src/slog_basic_test.dart
@@ -6,6 +6,7 @@
 // https://opensource.org/licenses/MIT.
 
 import 'package:slog_basic/slog_basic.dart';
+import 'package:slog_basic/src/printers.dart';
 // ignore_for_file: prefer_const_constructors
 import 'package:test/test.dart';
 
@@ -13,6 +14,11 @@ void main() {
   group('Basic Interceptors', () {
     test('Print Interceptor can be instantiated', () {
       expect(PrintInterceptor(), isNotNull);
+    });
+  });
+  group('Basic Printers', () {
+    test('Json Printer can be instantiated', () {
+      expect(JsonPrinter(), isNotNull);
     });
   });
 }