+++ packages/slog_core/lib/src/model/channel.dart
@@ -1,5 +1,8 @@
+///A log channel
 class SLogChannel {
+  ///A log channel
   const SLogChannel({required this.name});
 
+  ///The name of the log channel also functions as a channel key
   final String name;
 }