+++ packages/slog_core/lib/src/slogger.dart
@@ -6,18 +6,31 @@ import 'package:meta/meta.dart';
 import 'package:slog_core/src/errors.dart';
 import 'package:slog_core/src/model.dart';
 
+///The base logger class
 abstract class SLogger {
+  ///The base logger class
   SLogger({required SLogOptions options}) {
     _options = options;
   }
 
   late final SLogOptions _options;
 
+  ///Get the current options that the logger is using.
+  ///
+  ///Useful for mixins or when extended.
   @nonVirtual
   SLogOptions get options => _options;
 
   final Map<String, SLogChannel> _channels = {};
 
+  ///Get the current options that the logger is using.
+  ///
+  ///Useful for mixins or when extended.
+  List<SLogChannel> get channels => _channels.values.toList();
+
+  ///Register a new channel to the logger
+  ///
+  ///Throws [LogChannelExists]
   @nonVirtual
   void registerChannel(String channel) {
     if (_channels.containsKey(channel)) {
@@ -27,6 +40,9 @@ abstract class SLogger {
     _channels.addAll({channel: SLogChannel(name: channel)});
   }
 
+  ///Unregister a channel from the logger
+  ///
+  ///Throws [LogChannelDoesNotExist] if [SLogOptions.handleMissingChannels] is `false`
   @nonVirtual
   void unregisterChannel(String channel) {
     if (_channels.containsKey(channel)) {
@@ -38,8 +54,7 @@ abstract class SLogger {
     }
   }
 
-  List<SLogChannel> get channels => _channels.values.toList();
-
+  ///Used to run the interceptors and printers
   @protected
   @mustCallSuper
   Future<void> requestLog(LogRequest request) async {