+++ packages/slog_core/CHANGELOG.md
@@ -1,3 +1,5 @@
 # 0.1.0+1
 
 - feat: initial commit 🎉
+- feat: added log interceptor & printer
+- feat: added default logger